refactor(cli): move logic for reading artifact to helper fn
